feat(table-container): displaying data with tableContainer HOC
